+++ Domain/Interfaces/Repository/IUserRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IUserRepository
     {
-        Task<(bool Success, string Error, long UserId)> CreateUserAsync(string gamertag, string email, string password);
+        Task<(bool Success, string Error, int UserId)> CreateUserAsync(string gamertag, string email, string password);
         Task<UserInfo> FindByGamertagAsync(string gamertag);
         Task<bool> CheckPasswordAsync(UserInfo user, string password);
         Task<IList<string>> GetRolesAsync(UserInfo user);
